+++ src/ladb_opencutlist/ruby/worker/cutlist/cutlist_export_worker.rb
@@ -171,6 +171,7 @@ module Ladb::OpenCutList
               LengthWrapper.new(part.def.size.width),
               LengthWrapper.new(part.def.size.thickness),
               AreaWrapper.new(part.def.final_area),
+              StringWrapper.new(Plugin.instance.get_i18n_string("tab.materials.type_#{group.material_type}")),
               StringWrapper.new(group.material_display_name),
               ArrayWrapper.new(part.entity_names.map(&:first)),
               StringWrapper.new(part.description),
@@ -243,6 +244,7 @@ module Ladb::OpenCutList
                   LengthWrapper.new(part.def.size.width),
                   LengthWrapper.new(part.def.size.thickness),
                   AreaWrapper.new(part.def.final_area),
+                  StringWrapper.new(Plugin.instance.get_i18n_string("tab.materials.type_#{group.material_type}")),
                   StringWrapper.new(group.material_display_name),
                   StringWrapper.new(part.description),
                   ArrayWrapper.new(part.tags),
@@ -371,6 +373,7 @@ module Ladb::OpenCutList
       bbox_width,
       bbox_thickness,
       final_area,
+      material_type,
       material_name,
       entity_names,
       description,
@@ -392,6 +395,7 @@ module Ladb::OpenCutList
       @bbox_width = bbox_width
       @bbox_thickness = bbox_thickness
       @final_area = final_area
+      @material_type = material_type
       @material_name = material_name
       @entity_names = entity_names
       @description = description
@@ -421,6 +425,7 @@ module Ladb::OpenCutList
       bbox_width,
       bbox_thickness,
       final_area,
+      material_type,
       material_name,
       description,
       tags,
@@ -443,6 +448,7 @@ module Ladb::OpenCutList
       @bbox_width = bbox_width
       @bbox_thickness = bbox_thickness
       @final_area = final_area
+      @material_type = material_type
       @material_name = material_name
       @description = description
       @tags = tags